Frontend/fix: Dynamic feature fix for gullak
