+++ Frontend/android-native/app/build.gradle
@@ -357,6 +357,8 @@ dependencies {
     implementation 'com.github.freshworks:freshchat-android:6.0.0'
     api(platform("com.squareup.okhttp3:okhttp-bom:4.12.0"))
     api("com.squareup.okhttp3:okhttp")
+    compileOnly 'com.facebook.react:react-android:0.72.3'
+
 }
 
 def tmpSource = "/tmp/juspay"
