--- Frontend/android-native/dynamicfeature/src/main/java/in/juspay/mobility/dynamicfeature/MainApplication.java
@@ -1,60 +1,0 @@
-package in.juspay.mobility.dynamicfeature;
-
-import android.app.Application;
-import com.facebook.react.BuildConfig;
-import com.facebook.react.PackageList;
-import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactNativeHost;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.defaults.DefaultReactNativeHost;
-import com.facebook.soloader.SoLoader;
-import com.finternet.sdk.MyEventPackage;
-import java.util.List;
-
-public class MainApplication extends Application implements ReactApplication {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        SoLoader.init(this, false);
-    }
-
-
-    @Override
-    public ReactNativeHost getReactNativeHost() {
-        return createReactNativeHost();
-    }
-
-
-    public ReactNativeHost createReactNativeHost() {
-        return new DefaultReactNativeHost(this) {
-            @Override
-            public boolean getUseDeveloperSupport() {
-                return BuildConfig.DEBUG;
-            }
-
-            @Override
-            protected List<ReactPackage> getPackages() {
-                @SuppressWarnings("UnnecessaryLocalVariable")
-                List<ReactPackage> packages = new PackageList(this).getPackages();
-                packages.add(new MyEventPackage());
-                return packages;
-            }
-
-            @Override
-            protected String getJSMainModuleName() {
-                return "index";
-            }
-
-            @Override
-            protected boolean isNewArchEnabled() {
-                return false;
-            }
-
-            @Override
-            protected Boolean isHermesEnabled() {
-                return false;
-            }
-        };
-    }
-}
