+++ Frontend/android-native/mobility-driver/build.gradle
@@ -22,6 +22,7 @@ android {
         qa {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            matchingFallbacks = ['debug']
         }
         debug {
             minifyEnabled false
@@ -66,4 +67,5 @@ dependencies {
         exclude group: 'co.hyperverge', module: 'face-detection-preview-frame'
     }
     implementation 'com.google.code.gson:gson:2.8.8'
+    implementation 'com.facebook.react:react-android:0.72.3'
 }
