+++ Frontend/android-native/mobility-driver/src/main/AndroidManifest.xml
@@ -28,6 +28,11 @@
     <uses-permission android:name="android.Manifest.permission.QUERY_ALL_PACKAGES"/>
     <uses-permission android:name="android.permission.REQUEST_DELETE_PACKAGES"/>
 
+
+    <application
+        android:name=".MainApplication">
+    </application>
+
     <queries>
         <intent>
             <action android:name="android.intent.action.MAIN" />
